refactored for code modularity
