+++ src/data_providers.rs
@@ -1,6 +1,7 @@
 pub mod binance;
 pub mod bybit;
 
+// data types
 #[derive(Debug, Clone, Copy, Default)]
 pub struct Order {
     pub price: f32,
@@ -139,4 +140,132 @@ pub trait DataProvider {
     fn get_orderbook(&self, symbol: &str) -> Result<Depth, Box<dyn std::error::Error>>;
 
     fn get_trades(&self, symbol: &str) -> Result<Vec<Trade>, Box<dyn std::error::Error>>;
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct TickMultiplier(pub u16);
+
+impl std::fmt::Display for TickMultiplier {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}x", self.0)
+    }
+}
+
+impl TickMultiplier {
+    pub fn multiply_with_min_tick_size(&self, min_tick_size: f32) -> f32 {
+        self.0 as f32 * min_tick_size
+    }
+}
+
+// connection types
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum Exchange {
+    BinanceFutures,
+    BybitLinear,
+}
+
+impl std::fmt::Display for Exchange {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                Exchange::BinanceFutures => "Binance Futures",
+                Exchange::BybitLinear => "Bybit Linear",
+            }
+        )
+    }
+}
+impl Exchange {
+    pub const ALL: [Exchange; 2] = [Exchange::BinanceFutures, Exchange::BybitLinear];
+}
+
+impl std::fmt::Display for Ticker {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                Ticker::BTCUSDT => "BTCUSDT",
+                Ticker::ETHUSDT => "ETHUSDT",
+                Ticker::SOLUSDT => "SOLUSDT",
+                Ticker::LTCUSDT => "LTCUSDT",
+            }
+        )
+    }
+}
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum Ticker {
+    BTCUSDT,
+    ETHUSDT,
+    SOLUSDT,
+    LTCUSDT,
+}
+impl Ticker {
+    pub const ALL: [Ticker; 4] = [Ticker::BTCUSDT, Ticker::ETHUSDT, Ticker::SOLUSDT, Ticker::LTCUSDT];
+}
+
+impl std::fmt::Display for Timeframe {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                Timeframe::M1 => "1m",
+                Timeframe::M3 => "3m",
+                Timeframe::M5 => "5m",
+                Timeframe::M15 => "15m",
+                Timeframe::M30 => "30m",
+            }
+        )
+    }
+}
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum Timeframe {
+    M1,
+    M3,
+    M5,
+    M15,
+    M30,
+}
+impl Timeframe {
+    pub const ALL: [Timeframe; 5] = [Timeframe::M1, Timeframe::M3, Timeframe::M5, Timeframe::M15, Timeframe::M30];
+}
+
+#[derive(Debug)]
+pub enum BinanceWsState {
+    Connected(binance::market_data::Connection),
+    Disconnected,
+}
+impl Default for BinanceWsState {
+    fn default() -> Self {
+        Self::Disconnected
+    }
+}
+
+#[derive(Debug)]
+pub enum BybitWsState {
+    Connected(bybit::market_data::Connection),
+    Disconnected,
+}
+impl Default for BybitWsState {
+    fn default() -> Self {
+        Self::Disconnected
+    }
+}
+
+pub enum UserWsState {
+    Connected(binance::user_data::Connection),
+    Disconnected,
+}
+impl Default for UserWsState {
+    fn default() -> Self {
+        Self::Disconnected
+    }
+}
+
+#[derive(Debug, Clone)]
+pub enum MarketEvents {
+    Binance(binance::market_data::Event),
+    Bybit(bybit::market_data::Event),
 }
